Extract common collection logic
Not 100% sure of using callback for counting, but it does highlight
what is the part that is different. Particularly, I'm not super keen on
giving the number of people as an input. But since it's all I have
right now it seems good enough .Better than passing the full group.